Web.Authenticate.OAuth

Data types

data OAuth

Data type for OAuth client (consumer).

The constructor for this data type is not exposed. Instead, you should use the def method or newOAuth function to retrieve a default instance, and then use the records below to make modifications. This approach allows us to add configuration options without breaking backwards compatibility.

def :: Default a => a

The default value for this type.

newOAuth :: OAuth

Default value for OAuth datatype. You must specify at least oauthServerName, URIs and Tokens.

oauthServerName :: OAuth -> String

Service name (default: "")

oauthRequestUri :: OAuth -> String

URI to request temporary credential (default: ""). You MUST specify if you use getTemporaryCredential', getTemporaryCredentialProxy or getTemporaryCredential; otherwise you can just leave this empty.

oauthAccessTokenUri :: OAuth -> String

Uri to obtain access token (default: ""). You MUST specify if you use getAcessToken or getAccessToken' or getAccessTokenWith; otherwise you can just leave this empty.

oauthAuthorizeUri :: OAuth -> String

Uri to authorize (default: ""). You MUST specify if you use authorizeUrl or authorizeZUrl'; otherwise you can just leave this empty.

oauthSignatureMethod :: OAuth -> SignMethod

Signature Method (default: HMACSHA1)

oauthConsumerKey :: OAuth -> ByteString

Consumer key (You MUST specify)

oauthConsumerSecret :: OAuth -> ByteString

Consumer Secret (You MUST specify)

oauthCallback :: OAuth -> Maybe ByteString

Callback uri to redirect after authentication (default: Nothing)

oauthRealm :: OAuth -> Maybe ByteString

Optional authorization realm (default: Nothing)

oauthVersion :: OAuth -> OAuthVersion

OAuth spec version (default: OAuth10a)

data OAuthVersion

Constructors

OAuth10
OAuth protocol ver 1.0 (no oauth_verifier; differs from RFC 5849).

OAuth10a
OAuth protocol ver 1.0a. This corresponds to community's 1.0a spec and RFC 5849.

data SignMethod

Data type for signature method.

Constructors

PLAINTEXT

HMACSHA1

RSASHA1 PrivateKey

newtype Credential

Data type for credential.

Constructors

Credential

Fields

unCredential :: [(ByteString, ByteString)]

newtype OAuthException

Constructors

OAuthException String

Access token request

data AccessTokenRequest

Data type for getAccessTokenWith method.

You can create values of this type using defaultAccessTokenRequest.

Since 1.5.1

defaultAccessTokenRequest :: OAuth -> Credential -> Manager -> AccessTokenRequest

Create a value of type AccessTokenRequest with default values filled in.

Note that this is a settings type. More information on usage can be found at: http://www.yesodweb.com/book/settings-types.

Since 1.5.1

accessTokenAddAuth :: AccessTokenRequest -> ByteString -> Credential -> Request -> Request

add auth hook.

Default: addAuthHeader

Since 1.5.1

accessTokenRequestHook :: AccessTokenRequest -> Request -> Request

Request Hook.

Default: id

Since 1.5.1

accessTokenOAuth :: AccessTokenRequest -> OAuth

OAuth Application

Since 1.5.1

accessTokenTemporaryCredential :: AccessTokenRequest -> Credential

Temporary Credential (with oauth_verifier if >= 1.0a)

Since 1.5.1

accessTokenManager :: AccessTokenRequest -> Manager

Manager

Since 1.5.1

Operations for credentials

newCredential

Arguments

:: ByteString
value for oauth_token

-> ByteString
value for oauth_token_secret

-> Credential

Convenient function to create Credential with OAuth Token and Token Secret.

emptyCredential :: Credential

Empty credential.

insert

Arguments

:: ByteString
Parameter Name

-> ByteString
Value

-> Credential
Credential

-> Credential
Result

Insert an oauth parameter into given Credential.

delete

Arguments

:: ByteString
Parameter name

-> Credential
Credential

-> Credential
Result

Remove an oauth parameter for key from given Credential.

inserts :: [(ByteString, ByteString)] -> Credential -> Credential

Convenient method for inserting multiple parameters into credential.

injectVerifier :: ByteString -> Credential -> Credential

Insert oauth-verifier on a Credential.

Signature

signOAuth

Arguments

:: MonadIO m

=> OAuth
OAuth Application

-> Credential
Credential

-> Request
Original Request

-> m Request
Signed OAuth Request

Add OAuth headers & sign to Request.

genSign :: MonadIO m => OAuth -> Credential -> Request -> m ByteString

Generate OAuth signature. Used by signOAuth.

checkOAuth :: MonadIO m => OAuth -> Credential -> Request -> ExceptT OAuthException m Request

Test existing OAuth signature. Since 1.5.2

Url & operation for authentication

Temporary credentials

getTemporaryCredential

Arguments

:: MonadIO m

=> OAuth
OAuth Application

-> Manager

-> m Credential
Temporary Credential (Request Token & Secret).

Get temporary credential for requesting acces token.

getTemporaryCredentialWithScope

Arguments

:: MonadIO m

=> ByteString
Scope parameter string

-> OAuth
OAuth Application

-> Manager

-> m Credential
Temporay Credential (Request Token & Secret).

Get temporary credential for requesting access token with Scope parameter.

getTemporaryCredentialProxy

Arguments

:: MonadIO m

=> Maybe Proxy
Proxy

-> OAuth
OAuth Application

-> Manager

-> m Credential
Temporary Credential (Request Token & Secret).

Get temporary credential for requesting access token via the proxy.

getTemporaryCredential'

Arguments

:: MonadIO m

=> (Request -> Request)
Request Hook

-> OAuth
OAuth Application

-> Manager

-> m Credential
Temporary Credential (Request Token & Secret).

Authorization URL

authorizeUrl

Arguments

:: OAuth
OAuth Application

-> Credential
Temporary Credential (Request Token & Secret)

-> String
URL to authorize

URL to obtain OAuth verifier.

authorizeUrl'

Arguments

:: (OAuth -> Credential -> SimpleQuery)

-> OAuth
OAuth Application

-> Credential
Temporary Credential (Request Token & Secret)

-> String
URL to authorize

Convert OAuth and Credential to URL to authorize. This takes function to choice parameter to pass to the server other than oauth_callback or oauth_token.

Attaching auth to requests

addAuthBody :: a -> Credential -> Request -> Request

Place the authentication information in a URL encoded body instead of the Authorization header.

Note that the first parameter is used for realm in addAuthHeader, and this function needs the same type. The parameter, however, is unused.

Since 1.5.1

Finishing authentication

getAccessToken

Arguments

:: MonadIO m

=> OAuth
OAuth Application

-> Credential
Temporary Credential (with oauth_verifier if >= 1.0a)

-> Manager

-> m Credential
Token Credential (Access Token & Secret)

Get Access token.

getAccessTokenProxy

Arguments

:: MonadIO m

=> Maybe Proxy
Proxy

-> OAuth
OAuth Application

-> Credential
Temporary Credential (with oauth_verifier if >= 1.0a)

-> Manager

-> m Credential
Token Credential (Access Token & Secret)

Get Access token via the proxy.

getTokenCredential

Arguments

:: MonadIO m

=> OAuth
OAuth Application

-> Credential
Temporary Credential (with oauth_verifier if >= 1.0a)

-> Manager

-> m Credential
Token Credential (Access Token & Secret)

Get Access token.

getTokenCredentialProxy

Arguments

:: MonadIO m

=> Maybe Proxy
Proxy

-> OAuth
OAuth Application

-> Credential
Temporary Credential (with oauth_verifier if >= 1.0a)

-> Manager

-> m Credential
Token Credential (Access Token & Secret)

Get Access token via the proxy.

getAccessToken'

Arguments

:: MonadIO m

=> (Request -> Request)
Request Hook

-> OAuth
OAuth Application

-> Credential
Temporary Credential (with oauth_verifier if >= 1.0a)

-> Manager

-> m Credential
Token Credential (Access Token & Secret)

getAccessTokenWith

Arguments

:: MonadIO m

=> AccessTokenRequest
extensible parameters

-> m (Either (Response ByteString) Credential)
Token Credential (Access Token & Secret) or the conduit response on failures

Utility Methods

paramEncode :: ByteString -> ByteString

Encode a string using the percent encoding method for OAuth.

addScope :: ByteString -> Request -> Request

addMaybeProxy :: Maybe Proxy -> Request -> Request