Web.Authenticate.OAuth.IO

Description

Deprecated: This module is deprecated; rewrite your code using MonadResource

This Module provides interface for the instance of MonadIO instead of MonadIO. What this module do is just adding withManager or runResourceT.

Documentation

module Web.Authenticate.OAuth

getAccessToken

Arguments

:: MonadIO m

=> OAuth
OAuth Application

-> Credential
Temporary Credential with oauth_verifier

-> m Credential
Token Credential (Access Token & Secret)

Get Access token.

getTemporaryCredential

Arguments

:: MonadIO m

=> OAuth
OAuth Application

-> m Credential
Temporary Credential (Request Token & Secret).

Get temporary credential for requesting acces token.

getTemporaryCredentialWithScope

Arguments

:: MonadIO m

=> ByteString
Scope parameter string

-> OAuth
OAuth Application

-> m Credential
Temporay Credential (Request Token & Secret).

Get temporary credential for requesting access token with Scope parameter.

getTemporaryCredentialProxy

Arguments

:: MonadIO m

=> Maybe Proxy
Proxy

-> OAuth
OAuth Application

-> m Credential
Temporary Credential (Request Token & Secret).

Get temporary credential for requesting access token via the proxy.

getTemporaryCredential'

Arguments

:: MonadIO m

=> (Request -> Request)
Request Hook

-> OAuth
OAuth Application

-> m Credential
Temporary Credential (Request Token & Secret).

getTokenCredential

Arguments

:: MonadIO m

=> OAuth
OAuth Application

-> Credential
Temporary Credential with oauth_verifier

-> m Credential
Token Credential (Access Token & Secret)

Get Access token.

getAccessTokenProxy

Arguments

:: MonadIO m

=> Maybe Proxy
Proxy

-> OAuth
OAuth Application

-> Credential
Temporary Credential with oauth_verifier

-> m Credential
Token Credential (Access Token & Secret)

Get Access token via the proxy.

getTokenCredentialProxy

Arguments

:: MonadIO m

=> Maybe Proxy
Proxy

-> OAuth
OAuth Application

-> Credential
Temporary Credential with oauth_verifier

-> m Credential
Token Credential (Access Token & Secret)

Get Access token via the proxy.

getAccessToken'

Arguments

:: MonadIO m

=> (Request -> Request)
Request Hook

-> OAuth
OAuth Application

-> Credential
Temporary Credential with oauth_verifier

-> m Credential
Token Credential (Access Token & Secret)